+++ web_v2/handlers/get_history_base.go
@@ -23,6 +23,8 @@ type getHistory struct {
 	Balance *history2.Balance
 	Asset   *core2.Asset
 
+	BalancesIDs []uint64
+
 	AssetsQ   core2.AssetsQ
 	EffectsQ  history2.ParticipantEffectsQ
 	AccountsQ history2.AccountsQ
@@ -41,7 +43,6 @@ func newHistoryHandler(r *http.Request) getHistory {
 	}
 
 	return handler
-
 }
 
 func (h *getHistory) prepare(w http.ResponseWriter, r *http.Request) (*requests.GetHistory, bool) {
@@ -70,8 +71,63 @@ func (h *getHistory) prepare(w http.ResponseWriter, r *http.Request) (*requests.
 		}
 	}
 
-	if request.Filters.Balance != nil {
-		h.Balance, err = h.BalanceQ.GetByAddress(*request.Filters.Balance)
+	if len(request.Filters.Balance) > 1 {
+		balances, err := h.BalanceQ.SelectByAddress(request.Filters.Balance...)
+		if err != nil {
+			ctx.Log(r).WithError(err).Error("failed to get balances", logan.F{
+				"balances": request.Filters.Balance,
+			})
+			ape.RenderErr(w, problems.InternalError())
+			return nil, false
+		}
+
+		loadedBalances := make(map[string]bool)
+		differentBalances := make([]string, 0, len(balances))
+		for idx, balance := range balances {
+			loadedBalances[balance.Address] = true
+			h.BalancesIDs = append(h.BalancesIDs, balances[idx].ID)
+			if balance.AccountID != balances[0].AccountID {
+				differentBalances = append(differentBalances, balance.Address)
+			}
+		}
+		if len(differentBalances) != 0 {
+			ctx.Log(r).Error("balances must belong to the same account", logan.F{
+				"balance": differentBalances,
+			})
+			ape.RenderErr(w, problems.Conflict())
+			return nil, false
+		}
+
+		if len(balances) != len(request.Filters.Balance) {
+			notFoundBalances := make([]string, 0, len(balances))
+			for _, balance := range request.Filters.Balance {
+				if _, loaded := loadedBalances[balance]; !loaded {
+					notFoundBalances = append(notFoundBalances, balance)
+				}
+			}
+
+			if len(notFoundBalances) > 0 {
+				ctx.Log(r).Error("balance not found", logan.F{
+					"balance": notFoundBalances,
+				})
+				ape.RenderErr(w, problems.NotFound())
+			} else {
+				ctx.Log(r).Error("balance duplicate")
+				ape.RenderErr(w, problems.Conflict())
+			}
+
+			return nil, false
+		}
+
+		if !h.ensureAllowed(w, r) {
+			return nil, false
+		}
+
+		return request, true
+	}
+
+	if len(request.Filters.Balance) == 1 {
+		h.Balance, err = h.BalanceQ.GetByAddress(request.Filters.Balance[0])
 		if err != nil {
 			ctx.Log(r).WithError(err).Error("failed to get balance", logan.F{
 				"balance_address": request.Filters.Balance,
@@ -86,6 +142,8 @@ func (h *getHistory) prepare(w http.ResponseWriter, r *http.Request) (*requests.
 			})...)
 			return nil, false
 		}
+
+		h.BalancesIDs = append(h.BalancesIDs, h.Balance.ID)
 	}
 
 	if request.Filters.Asset != nil {
@@ -124,8 +182,8 @@ func (h *getHistory) ApplyFilters(request *requests.GetHistory,
 		q = q.ForAccount(h.Account.ID)
 	}
 
-	if h.Balance != nil {
-		q = q.ForBalance(h.Balance.ID)
+	if len(h.BalancesIDs) > 0 {
+		q = q.ForBalance(h.BalancesIDs...)
 	}
 
 	if h.Asset != nil {