+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * Reviewable request for create, update and remove data 
 * Ability to set custom rules and actions in permissions
 * Filter by effect in get history endpoint
+* Fitler by `created_before`, `created_after` timestamps for reviewable requests
+* Filter by `all_tasks`, `all_tasks_any_of`, `all_tasks_not_set` for reviewable requests
 
 ### Changed
 
@@ -16,6 +18,7 @@
 
 * `/v3/order_book/{id}` now uses renders numbers with trailing-digits-count-based dot
 * Match participant in manage offer
+* Error on '/v3/sales/{id}/relationships/participation'
 
 ## 3.9.1